Split get_info into two functions
This is part of a refactoring effort for get_info, has_info, and
save_info.  Part of the reason their interface sucks so much is
because they try to do too much. If the functions are specialized by
the kind of information they're looking for, everything makes much
more sense. Too bad I didn't realize this earlier.